test: get tests working after refactor
