+++ frontend/routes/auth/auth.spec.tsx
@@ -2,6 +2,8 @@ import { render, screen } from '@testing-library/react'
 import { Auth } from '.'
 import { MemoryRouter } from 'react-router-dom'
 import locators from '../../components/locators'
+import { JsonRPCProvider } from '../../contexts/json-rpc/json-rpc-provider'
+import { mockClient } from '../../test-helpers/mock-client'
 
 jest.mock('react-router-dom', () => ({
   ...(jest.requireActual('react-router-dom') as any),
@@ -18,14 +20,20 @@ jest.mock('../../components/transaction-modal', () => ({
 
 describe('Auth', () => {
   it('renders outlet, header and navbar', () => {
+    mockClient()
     render(
-      <MemoryRouter>
-        <Auth />
-      </MemoryRouter>
+      <JsonRPCProvider>
+        <MemoryRouter>
+          <Auth />
+        </MemoryRouter>
+      </JsonRPCProvider>
     )
     expect(screen.getByTestId(locators.navBar)).toBeInTheDocument()
     expect(screen.getByTestId('outlet')).toBeInTheDocument()
     expect(screen.getByTestId('connection-modal')).toBeInTheDocument()
     expect(screen.getByTestId('transaction-modal')).toBeInTheDocument()
+  })
+  it('loads the users wallets', () => {
+    expect(false).toBe(true)
   })
 })
