+++ frontend/stores/wallets.ts
@@ -9,7 +9,6 @@ export interface Wallet {
 
 export interface Key {
   index: number
-  metadata: Metadata[]
   name: string
   publicKey: string
 }
@@ -66,6 +65,7 @@ export const useWalletStore = create<WalletsStore>()((set, get) => ({
           return { name: w, keys }
         })
       )
+      console.log(res)
       set({ wallets: res })
     } catch (e) {
       set({ error: e?.toString() || 'Something went wrong' })
